+++ addon/components/ivy-tabs.js
@@ -36,6 +36,7 @@ export default Ember.Component.extend({
    */
   registerTabList: function(tabList) {
     this.set('tabList', tabList);
+    Ember.run.once(this, this._selectTabByIndex);
   },
 
   /**
@@ -70,5 +71,11 @@ export default Ember.Component.extend({
 
   _initTabPanels: function() {
     this.set('tabPanels', Ember.A());
+  },
+
+  _selectTabByIndex: function() {
+    var selectedIndex = this.get('selected-index');
+    if (Ember.isNone(selectedIndex)) { selectedIndex = 0; }
+    this.get('tabList').selectTabByIndex(selectedIndex);
   }
 });